backend/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-Cors==4.0.0
python-socketio==5.11.0
pytest==8.2.0
eventlet==0.35.2